--- iOS_Wegg/iOS_Wegg/Screen/Map/MapSearch/ViewController/MapSearchResultViewController.swift
@@ -1,18 +1,0 @@
-//
-//  MapSearchResultViewController.swift
-//  iOS_Wegg
-//
-//  Created by jaewon Lee on 2/4/25.
-//
-
-import UIKit
-
-class MapSearchResultViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        view = MapSearchResultView()
-    }
-
-}
