+++ iOS_Wegg/iOS_Wegg/Screen/Map/MapSearch/ViewController/MapSearchViewController.swift
@@ -12,48 +12,29 @@ class MapSearchViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
+        
         view = mapSearchView
-        // Do any additional setup after loading the view.
     }
     
     lazy var mapSearchView = MapSearchView().then {
-        $0.gestureDelegte = self
+        $0.searchBarView.delegate = self
     }
 
 }
 
 // MARK: - Delegate Extenstion
 
-extension MapSearchViewController:
-    MapSearchViewGestureDelegate,
-    UITextFieldDelegate {
+extension MapSearchViewController: MapSearchBarDelegate {
     
-    // MARK: - MapSearchViewGestureDelegate
+    func didTapSearchBox() {
+        print("didTapSearchBox")
+    }
     
     func didTapSearchBackButton() {
         navigationController?.popViewController(animated: true)
     }
     
-    func didTapSearchButton() {
-        print("검색 하기")
-    }
-    
-    // MARK: - UITextFieldDelegate
-    
-    // 탭하여 편집을 시작할 때 호출
-    func textFieldDidBeginEditing(_ textField: UITextField) {
-        
-    }
-    
-    // 엔터(Return 키) 누를 때 호출되는 함수(검색 결과 처리)
-    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
-        guard let query = textField.text,
-              !query.isEmpty else {
-            return false
-        }
-        
-        print("입력된 값: \(query)")
-        
-        return true
+    func didSearch(query: String?) {
+        print("Search button tapped with query: \(query ?? "empty")")
     }
 }
